pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-functions-agents"
version = "0.7.1"
description = "A markdown-first programming model for building AI agents on Azure Functions with the GitHub Copilot SDK."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "azure-functions>=1.24.0",
    "github-copilot-sdk>=0.2.1",
    "pydantic>=2.0.0",
    "python-frontmatter>=1.1.0",
    "azurefunctions-extensions-http-fastapi",
    "aiohttp>=3.9.0",
    "azure-identity>=1.15.0",
]

[project.optional-dependencies]
connectors = [
    "azure-functions-connectors @ https://github.com/anthonychu/azure-functions-connectors-python/releases/download/v0.1.0b2/azure_functions_connectors-0.1.0b2-py3-none-any.whl",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
azure_functions_agents = ["public/**"]
